+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ elements
 * EspDialog - dismiss dialog but only if dialog config allow it
 * EspPermissionDialog - stabilize for pre marshmallow
 * EspPermissionDialog - use current localisation for button text
+* EspApplicationIsNotRespondingDialog - close anr dialog if shown
 * (planed) EspAlertDialog - change prepared ids for standard alert dialog
 * (planed) EspSupportAlertDialog - prepared ids for alert dialog from support package
 
@@ -15,7 +16,7 @@ tools
 
 test base
 
-* EspressoTestBase - base for all test cases
+* EspressoTestBase - base for all espresso test cases (finish resumed activities, close anr dialogs, take screenshot when test fails)
 * EspressoTestCase - base for default espresso tests
 * EspressoIntentTestCase - base test case when for mocking intents
 * EspCloseAllActivitiesFunction - use it to ensure that all started activities are finished